Add support for futures MSI creation
The introduction of proper semver support with bumpversion will
break the creation of MSI packages. The reason is that
distutils.version.StrictVersion does not support semantic versioning
for pre-releases. As a workaround, normalize the version to be
acceptable for StrictVersion. 

Note that the normalized version is identical to the semver except for
pre-releases (where '-dev' is replaced by 'a0').